# public/api-docs/swaggerDoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "#### WeTravel\n #### WeTravel is an application that will enable its Company Staff book their international travel and accommodation globally; easily and conveniently across all the locations/centers where the Company has its operations. \n\n #### While early travel tended to be slower, more dangerous, and more dominated by trade and migration, WeTravel uses cultural and technological advances to make travel become easier and more accessible.",
        "version": "1",
        "title": "WeTravel",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "vokeolomu@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/",
    "tags": [
        {
            "name": "Request",
            "description": "Access to WeTravel orders"
        },
        {
            "name": "User",
            "description": "Operations on WeTravel User"
        }
    ],
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/api/v1/request": {
            "post": {
                "tags": [
                    "Request"
                ],
                "summary": "Create a booking order",
                "description": "Returns a map of status codes to quantities",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Required fields",
                        "description": "Required field",
                        "schema": {
                            "type": "object",
                            "required": [
                                "userName"
                            ],
                            "properties": {
                                "price_offered": {
                                    "type": "integer"
                                },
                                "buyer": {
                                    "type": "string"
                                }
                            }
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Request"
                        }
                    },
                    "400": {
                        "description": "Invalid"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "signup",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@company.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/v1/auth/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Admin create a new user",
                "description": "Allows an admin to create a new user",
                "operationId": "CreateUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "firstName": {
                                    "type": "string",
                                    "example": "john"
                                },
                                "lastName": {
                                    "type": "string",
                                    "example": "doe"
                                },
                                "userName": {
                                    "type": "string",
                                    "example": "johnny"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@company.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password"
                                },
                                "phoneNumber": {
                                    "type": "string",
                                    "example": "09023232322"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
    "definitions": {
        "Request": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "price": {
                    "type": "integer",
                    "description": "price offered",
                    "example": "500,000.00"
                },
                "created_on": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "string",
                    "description": "Order Status",
                    "enum": [
                        "pending",
                        "accepted",
                        "rejected"
                    ],
                    "example": "pending"
                }
            },
            "xml": {
                "name": "Order"
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "token": {
                    "type": "integer",
                    "format": "int64"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@company.com"
                },
                "password": {
                    "type": "string"
                },
                "address": {
                    "type": "string",
                    "description": "Address of WeTravel User"
                }
            },
            "xml": {
                "name": "User"
            }
        },
        "externalDocs": {
            "description": "Find out more on WeTravel site",
            "url": "https://github.com/andela/phoenix-bn-backend"
        }
    }
}